Os/File: fix Linux File illegal close after failed open
If Os/Linux/File fails to open the file (receives fd -1), it will report
an error, but still save the fd and mode in its member variable. Even if
the user doesn't call close() on the file, the destructor of File will
still detect the mode and think the File is open, and attempt an illegal
close.

This patch fixes open() so that the fd and mode won't be saved if the
open failed.