Fix dictionary target dependencies
The dictonary target wasn't depending on the topology xml file,
so changes to the topology weren't causing the dictionary to rebuild.

The autocoding step also generates dictionaries as a side affect of
generating the topology autocode (might be worth considering if we can
somehow combine these steps so we don't have to run the topology autocoder
twice). This could sometimes run after the dictionary generation step,
leading to leftover dictionary files in the source tree. Adding a dependency
on the modules autocoded files (including the autocoded topology files)
ensures that this step always runs after the topology autocode step
and can properly clean up.

Also fix integration test was using old dictionary location but was
passing due to above bug.