Add the "Open in VS Code" badge to improve UX
The June 2021 release of Visual Studio Code includes the addition of the "Open in VS Code" badge to allow visitors to quickly open a GitHub repo in the VS Code open source editor.

With this badge, visitors can either use the recently launched VS Code Remote Repositories extension or clone into a Dev Container. recently launched or clone into a Dev Container.

By clicking on the badge, the visitor will be directed to a page on open.vscode.dev, where they can choose how to continue in VS Code; an example is available on the VS Code GitHub repo: microsoft/vscode.